Completed Problem 67. Also cleaned up the project a bit.
