--- src/SmallestMultiple.java
@@ -1,26 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-
-/**
- *
- * @author G
- */
-public class SmallestMultiple {
-    /*
-    2520 is the smallest number that can be divided by each of the numbers from 1 to 10 without any remainder.
-    What is the smallest positive number that is evenly divisible by all of the numbers from 1 to 20?
-    */
-    
-    public static void main(String[] args){
-        //Messed around one morning and figured it out:
-        
-        //The number has to be a prime factor of all primes between 1 and 20, so I used these then tweaked the numbers
-        //till I happened across it.
-        
-        //answer = (3*5*7*11*13*17*19)*(2*2*2*2*3) = 232792560
-    }
-    
-}
